third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbcursor-reused-expected.txt

This is a testharness.js-based test.
PASS IDBCursor is reused 
Harness: the test ran to completion.
